pyproject.toml
[tool.poetry]
name = "store-api-tdd"
version = "0.1.0"
description = ""
authors = ["dokidokiab <dokidokiabr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.2.2"
pydantic-settings = "^2.3.1"
motor = "^3.4.0"
pre-commit = "^3.7.1"
pytest-asyncio = "^0.23.7"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
fastapi = "^0.111.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
  "--strict-config",
  "--strict-markers",
  "--ignore=docs_src",
]
xfail_strict = true
junit_family = "xunit2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
